CP-30335: Add abstract class WarningWithMoreInfo and use it for two of the existing Warnings
